==> reports/1_training_data_cleaning.ipynb <==
import pandas as pd

# Training Data Cleaning

In this notebook I will perform the data cleaning from a more engineer point of view. I will check for consistency, analyze possible errors in the data (missing values, outliers) and drop duplicated information

dataset = pd.read_csv('../data/Reviews.csv')

Missing values in ProfileName and Summary but not in Text and Score that will be our top priority features

dataset.isnull().sum()

dataset.Score.value_counts()

### Helpfulness review

Helpfulness Numerator: Number of users who found the review helpful

Helpfulness Denominator: Number of users who indicated whether they found the review helpful or not

Considering this, possible feature:
- NotHelpfull instead of the helpfulness denominator
- Percentage of helpfulness

(dataset.HelpfulnessNumerator == dataset.HelpfulnessDenominator).sum() / dataset.shape[0]

(dataset.HelpfulnessDenominator < dataset.HelpfulnessNumerator).sum()

dataset[dataset.HelpfulnessDenominator < dataset.HelpfulnessNumerator]

There are two rows with inconsistencies in the data since the numerator is greater than the numerator

## Analyzing duplicates

3 possible subsets have been considerated for being a duplicate:

- All columns equal: there is none
- user and text the same: dropped those which had the same user and text, considering them the same review.
- user and time the same: since the time is in days, after a brief analysis we discover that there are genuine different reviews, we keep them.

For the process of which one keep, we considered to keep the last one and if they are both the same (because after a brief analysis we discovered only a few had different times in the same user and text), we would keep the one with more information in the helfulness columns. But after keeping only the text for the models, we decided to only keep the last that occured. 

One more thing to do for checking the consistency would be check if the score is the same in both reviews.

dataset.duplicated().sum()

text_user_duplicated = dataset[dataset.duplicated(subset=['UserId', 'Text'])]

time_user_duplicated = dataset[dataset.duplicated(subset=['UserId', 'Time'])]

### Converting date into datetime

dataset['date'] = pd.to_datetime(dataset['Time'], unit='s')

## Mini-checks of data consistency

## ProfileName and userId?

After analyzing the results, we have dropped the ProfileName. UserId is more consistent for tracking the user, since the ProfileName can change or be anonymous ('A customer'). It could be useful in a very far approach by trying to consider the gender, some demographic information or some analysis in possible change of behaviour when having an open name or not. But it's out of the scope of this analysis.

profname_by_userid = dataset.groupby('UserId').agg({'ProfileName': set})

(profname_by_userid.ProfileName.str.len() == 1).sum() / profname_by_userid.shape[0]

profname_by_userid[profname_by_userid.ProfileName.str.len() != 1].head()

### Text and summary

The same text doesn't have the same summary, so we have decided to drop the column. This summary isn't very good overall. However, using this or an alternative summary, could be useful maybe if the model that performs the summary + smaller model trained with summaries, is smaller than the current one and it also overcomes the overheat of having two models, then it could be used as an alternative version of the model.

text_by_summary = dataset.groupby('Text').agg({'Summary': set})

Maybe first approach using indistinctly one of them and see if we can get a better summary with one of the AI services

text_by_summary[text_by_summary.Summary.str.len() != 1].head()

## Function to wrap all the cleaning

Considerations:

- Duplicated rows considered by the same userid and text are dropped
- Id as index
- ProfileName and Summary columns drops
- Since the helpfulness columns are going to be discarded for the model and since they are an insificant number, we will keep the inconsistent rows.

def basic_dataset_cleaning(data):
    duplicated_mask = data.sort_values('Time').duplicated(subset=['UserId', 'Text'], keep='last')
    new_dataset = data.drop(duplicated_mask[duplicated_mask].index, axis=0)
    new_dataset.index = new_dataset['Id']
    new_dataset.drop(['ProfileName', 'Summary', 'Id'], axis=1, inplace=True)
    new_dataset['Time'] = pd.to_datetime(data['Time'], unit='s')
    return new_dataset

new_dataset = basic_dataset_cleaning(dataset)
new_dataset[new_dataset.HelpfulnessDenominator < new_dataset.HelpfulnessNumerator]

new_dataset.to_csv("../data/cleaned_data.csv")
