==> reports/5_comparing_pretraining_models.ipynb <==
import pandas as pd
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import f1_score

This notebook focuses on selecting techniques for the final pipeline. For this initial model approach, various techniques are evaluated through cross-validation using a basic Logistic Regression model with default settings:

- If one technique demonstrates significantly better performance, it will be incorporated into the pipeline.
- If no significant difference is observed, the simplest technique will be selected.

Data indexes are tracked because some data is stored as a sparse NumPy array. Although the split was performed using a seed, consistency across library versions or operating systems cannot be guaranteed.

The **f1_macro** metric is used for evaluation, as it provides a balanced assessment of false positives and false negatives, ensuring equal consideration for both categories.

df = pd.read_csv('processed_text_with_all.csv')
df = df.dropna()

train_index = pd.read_csv('train_index.csv').values[:, 0]
test_index = pd.read_csv('test_index.csv').values[:, 0]

X_train = df.Text.loc[train_index]
X_test = df.Text.loc[test_index]
y_train = df.Labels.loc[train_index]
y_test = df.Labels.loc[test_index]

# Text to vector

This notebook explores different algorithms to convert processed text into vectors: CountVectorizer, TF-IDF, and Word2Vec. The Word2Vec training process is handled in the Word2VecTraining notebook, and here, the pre-trained data will be loaded.

- **300 features** have been chosen for all models to balance memory usage and avoid overfitting, maintaining a reasonable ratio between rows and columns. The models will be compared based on these 300 features.
- For both CountVectorizer and TF-IDF, 1-3 n-grams will be trained. The 2-gram model will capture word pairs like "not" and other composite terms, while the 3-gram model will ensure no useful information is overlooked.

count_vect3 = CountVectorizer(ngram_range=(1, 3), max_features=300)
X_train_cv3 = count_vect3.fit_transform(X_train) 
# X_val_cv = count_vect.transform(X_val)
X_test_cv3 = count_vect3.transform(X_test)

count_vect2 = CountVectorizer(ngram_range=(1, 2), max_features=300)
X_train_cv2 = count_vect2.fit_transform(X_train) 
# X_val_cv = count_vect.transform(X_val)
X_test_cv2 = count_vect2.transform(X_test)

count_vect1 = CountVectorizer(ngram_range=(1, 1), max_features=300)
X_train_cv1 = count_vect1.fit_transform(X_train) 
# X_val_cv = count_vect.transform(X_val)
X_test_cv1 = count_vect1.transform(X_test)

count_vect2.vocabulary_ == count_vect3.vocabulary_

count_vect2.vocabulary_ == count_vect1.vocabulary_

tfidf3 = TfidfVectorizer(ngram_range=(1, 3), max_features=300)
X_train_tfidf3 = tfidf3.fit_transform(X_train)
# X_val_tfidf = tfidf.transform(X_val)
X_test_tfidf3 = tfidf3.transform(X_test)

tfidf2 = TfidfVectorizer(ngram_range=(1, 2), max_features=300)
X_train_tfidf2 = tfidf2.fit_transform(X_train)
# X_val_tfidf = tfidf.transform(X_val)
X_test_tfidf2 = tfidf2.transform(X_test)

tfidf1 = TfidfVectorizer(ngram_range=(1, 1), max_features=300)
X_train_tfidf1 = tfidf1.fit_transform(X_train)
# X_val_tfidf = tfidf.transform(X_val)
X_test_tfidf1 = tfidf1.transform(X_test)

tfidf3.vocabulary_ == tfidf2.vocabulary_

tfidf2.vocabulary_ == tfidf1.vocabulary_

X_train_w2v = pd.read_csv('w2v_window10_train_data.csv', index_col=0)
X_test_w2v = pd.read_csv('w2v_window10_test_data.csv', index_col=0)

train_data_from_models = {'count_vec_ngram1': {'x': X_train_cv1,
                                               'y': y_train},
                          'tfidf_ngram1': {'x': X_train_tfidf1,
                                           'y': y_train},
                          'count_vec_ngram2': {'x': X_train_cv2,
                                               'y': y_train},
                          'tfidf_ngram2': {'x': X_train_tfidf2,
                                           'y': y_train},
                          'word2vec': {'x': X_train_w2v,
                                       'y': y_train}}

from sklearn.model_selection import cross_val_score


def generate_cross_scoring_table_from_transf_data(train_data_dict, 
                                                  pred_model, scoring='f1_macro', cv=10):

    all_scores = [cross_val_score(pred_model, train_data['x'], 
                                  train_data['y'], cv=cv, scoring=scoring)
                  for train_data in train_data_dict.values()]

    return pd.DataFrame(all_scores, index=train_data_dict.keys())

lr = LogisticRegression()


result = generate_cross_scoring_table_from_transf_data(train_data_from_models, 
                                                       lr, scoring='f1_macro', cv=10)


result

The word2vec is clearly better than the rest, so we will keep the data and the model from the word2vec for the following steps.

import pandas as pd

a = """0	1	2	3	4	5	6	7	8	9
count_vec_ngram1	0.730720	0.744704	0.740865	0.744642	0.733056	0.746572	0.741397	0.748263	0.743674	0.738538
tfidf_ngram1	0.747909	0.759543	0.754347	0.762461	0.754868	0.759464	0.753718	0.760397	0.758010	0.751561
count_vec_ngram2	0.734709	0.746985	0.743778	0.748821	0.737387	0.747985	0.744234	0.749215	0.743730	0.741885
tfidf_ngram2	0.751512	0.761197	0.756324	0.765519	0.757312	0.760695	0.753588	0.762479	0.759185	0.753557
word2vec	0.819960	0.833779	0.826369	0.834525	0.830219	0.830770	0.827782	0.834607	0.831262	0.830440""".split("\n")


b = [a1.split("\t") for a1 in a]

columns = b[0]
columns
c = pd.DataFrame(b[1:])
c.index = c[0]
c.drop(0, axis=1, inplace=True)
c.columns = columns
c = c.astype(float)
c.mean(axis=1).idxmax()

# Handle imbalance labels

Simple random undersampling, random oversampling, and SMOTE will be tested to evaluate if these techniques improve the model's performance.

A different cross-validation technique will be used to ensure the model is evaluated on data that has not been sampled, avoiding any potential bias.

from imblearn.under_sampling import RandomUnderSampler
from imblearn.over_sampling import RandomOverSampler
from imblearn.over_sampling import SMOTE

X_us_res, y_us_res = RandomUnderSampler().fit_resample(X_train_w2v, y_train)

X_os_res, y_os_res = RandomOverSampler().fit_resample(X_train_w2v, y_train)

X_sm_res, y_sm_res = SMOTE().fit_resample(X_train_w2v, y_train)

imb_models = {'base': None,
              'undersampling': RandomUnderSampler(),
              'oversampling': RandomOverSampler(),
              'smote': SMOTE()}

from sklearn.model_selection import ShuffleSplit


def generate_imb_results_table(pred_model_class, imb_models, X_train_orig, y_train_orig):
    all_results = []
    for model in imb_models.values():
        ss = ShuffleSplit(n_splits=10, test_size=0.2, random_state=0)
        ind_results = []
        for train_index, test_index in ss.split(X_train_orig.values):
            x_train = X_train_orig.iloc[train_index]
            y_train = y_train_orig.iloc[train_index]
            x_test = X_train_orig.iloc[test_index]
            y_test = y_train_orig.iloc[test_index]

            if model:
                x_train_imb, y_train_imb = model.fit_resample(x_train, y_train)
            else:
                x_train_imb, y_train_imb = x_train, y_train
            pred_model = pred_model_class()
            pred_model.fit(x_train_imb, y_train_imb)
            y_pred = pred_model.predict(x_test)
            ind_results.append(f1_score(y_test, y_pred, average='macro'))
            
        all_results.append(ind_results)
    return pd.DataFrame(all_results, index=imb_models.keys())

imb_result = generate_imb_results_table(LogisticRegression, imb_models, X_train_w2v, y_train)

imb_result

from scipy.stats import ttest_ind, wilcoxon


def test_ditr_significantly_greater_than_another(data1, data2, test_type, alpha=0.05):

    print(f'----------------------- {test_type.capitalize()} -----------------------')

    if test_type == 'ttest':
        test_func = ttest_ind
    elif test_type == 'wilcoxon':
        test_func = wilcoxon
    else:
        raise ValueError(f'test_type must be ttest or wilcoxon not {test_type}')

    statistic, p_value = test_func(data1, data2, alternative='greater')
    
    if p_value < alpha:
        print("The mean of the first distribution is significantly" +
            " greater than the mean of the second one.")
    else:
        print("The mean of the first distribution is not significantly" +
            " greater than the mean of the second one.")
        
    return statistic, p_value

statistic, p_value = ttest_ind(imb_result.loc['base'],
                               imb_result.loc['oversampling'], 
                               alternative='greater')

test_ditr_significantly_greater_than_another(imb_result.loc['base'], 
                                             imb_result.loc['oversampling'], 
                                             'ttest')
    
test_ditr_significantly_greater_than_another(imb_result.loc['base'], 
                                             imb_result.loc['oversampling'], 
                                             'wilcoxon')

test_ditr_significantly_greater_than_another(imb_result.loc['base'], 
                                             imb_result.loc['undersampling'], 
                                             'ttest')
    
test_ditr_significantly_greater_than_another(imb_result.loc['base'], 
                                             imb_result.loc['undersampling'], 
                                             'wilcoxon')

test_ditr_significantly_greater_than_another(imb_result.loc['base'], 
                                             imb_result.loc['smote'], 
                                             'ttest')
    
test_ditr_significantly_greater_than_another(imb_result.loc['base'], 
                                             imb_result.loc['smote'], 
                                             'wilcoxon')

It seems like for this dataset, the imbalancing techniques worsen the results. Therefore, we are skipping them.
